+++ src/features/errorBoundary/index.ts
@@ -1,3 +1,1 @@
-import ErrorBoundary from './ErrorBoundary';
-
-export default ErrorBoundary;
+export { default } from './ErrorBoundary';
