+++ src/features/theme/index.ts
@@ -1,0 +1,2 @@
+export { ThemeProvider, ThemeContext } from './ThemeContext';
+export type { ThemeContextProps } from './ThemeContext';
