+++ src/shared/hooks/useLocalStorage/index.ts
@@ -1,3 +1,1 @@
-import { useLocalStorage } from './useLocalStorage';
-
-export default useLocalStorage;
+export { default } from './useLocalStorage';
