+++ src/shared/hooks/useTheme/useTheme.ts
@@ -1,0 +1,12 @@
+import { useContext } from 'react';
+import { ThemeContext, ThemeContextProps } from '../../../features/theme/ThemeContext';
+
+const useTheme = (): ThemeContextProps => {
+  const context = useContext(ThemeContext);
+  if (!context) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
+
+export default useTheme;
