+++ src/shared/ui/searchButton/index.ts
@@ -1,3 +1,1 @@
-import SearchButton from './SearchButton';
-
-export default SearchButton;
+export { default } from './SearchButton';
