+++ src/widgets/cardList/index.ts
@@ -1,3 +1,1 @@
-import CardList from './CardList';
-
-export default CardList;
+export { default } from './CardList';
