+++ src/widgets/header/index.ts
@@ -1,3 +1,1 @@
-import Header from './Header';
-
-export default Header;
+export { default } from './Header';
