+++ src/widgets/pagination/index.ts
@@ -1,3 +1,1 @@
-import Pagination from './Pagination';
-
-export default Pagination;
+export { default } from './Pagination';
