+++ src/widgets/searchBar/index.ts
@@ -1,3 +1,1 @@
-import SearchBar from './SearchBar';
-
-export default SearchBar;
+export { default } from './SearchBar';
